# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DailyScreen Lite",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/quotes/summary": {
      "get": {
        "description": "批量轻量行情摘要：队列列表的最新收盘价、行情日与日涨跌幅。\n\n只读取最近两个交易日，不为列表加载整段历史；缺失的股票不返回条目。",
        "operationId": "quote_summaries_api_quotes_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "securityIds",
            "required": true,
            "schema": {
              "title": "Securityids",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Quote Summaries Api Quotes Summary Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Quote Summaries",
        "tags": [
          "quotes"
        ]
      }
    },
    "/api/quotes/{security_id}": {
      "get": {
        "description": "股票日线：默认最近约 250 个交易日，可用 end 向前浏览更早历史。",
        "operationId": "quote_view_api_quotes__security_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "security_id",
            "required": true,
            "schema": {
              "title": "Security Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 250,
              "maximum": 2000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Quote View Api Quotes  Security Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Quote View",
        "tags": [
          "quotes"
        ]
      }
    },
    "/api/securities/{security_id}/refresh": {
      "post": {
        "description": "历史详情的单股手动更新：只更新这一只，不加入持续更新范围。",
        "operationId": "refresh_single_security_api_securities__security_id__refresh_post",
        "parameters": [
          {
            "in": "path",
            "name": "security_id",
            "required": true,
            "schema": {
              "title": "Security Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Refresh Single Security Api Securities  Security Id  Refresh Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Single Security",
        "tags": [
          "quotes"
        ]
      }
    },
...
  }
}